Add protective coding to check for existence of bridge multicast_snooping kernel parameter before trying to set it. Resolves [Issue 7](https://github.com/adjacentlink/python-etce/issues/7).

+++ etce/linux/platformimpl.py
@@ -140,8 +140,17 @@ class PlatformImpl(etce.platformimpl.PlatformImpl):
             self.runcommand('ip link set %s up' % interface)
             
         if enablemulticastsnooping:
-            with open('/sys/devices/virtual/net/%s/bridge/multicast_snooping' % bridgename, 'w') as sf:
-                sf.write('0')
+
+            if os.path.exists('/sys/devices/virtual/net/%s/bridge/multicast_snooping' % bridgename):
+
+                with open('/sys/devices/virtual/net/%s/bridge/multicast_snooping' % bridgename, 'w') as sf:
+    
+                    sf.write('0')
+            else:
+                warning = 'Warning: /sys/devices/virtual/net/%s/bridge/multicast_snooping does not exist, ' \
+                    'cannot disable multicast_snooping.' % bridgename
+
+                print(warning, file=sys.stderr)
 
 
     def bridgedown(self, bridgename, addifs):
